backend/pronunciation/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opensmile==2.5.0
python-multipart==0.0.9
pydantic==2.5.3
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
librosa==0.10.1
soundfile==0.12.1
speechrecognition==3.10.1
torch==2.1.2
transformers==4.37.0
piper-tts==1.2.0
websockets==12.0
